+++ elements/src/components.d.ts
@@ -65,8 +65,6 @@ export namespace Components {
     }
     interface PosTypeRouter {
     }
-    interface TestComponent {
-    }
 }
 export interface PosAppDocumentViewerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -295,12 +293,6 @@ declare global {
         prototype: HTMLPosTypeRouterElement;
         new (): HTMLPosTypeRouterElement;
     };
-    interface HTMLTestComponentElement extends Components.TestComponent, HTMLStencilElement {
-    }
-    var HTMLTestComponentElement: {
-        prototype: HTMLTestComponentElement;
-        new (): HTMLTestComponentElement;
-    };
     interface HTMLElementTagNameMap {
         "pos-app": HTMLPosAppElement;
         "pos-app-browser": HTMLPosAppBrowserElement;
@@ -327,7 +319,6 @@ declare global {
         "pos-subjects": HTMLPosSubjectsElement;
         "pos-type-badges": HTMLPosTypeBadgesElement;
         "pos-type-router": HTMLPosTypeRouterElement;
-        "test-component": HTMLTestComponentElement;
     }
 }
 declare namespace LocalJSX {
@@ -360,11 +351,19 @@ declare namespace LocalJSX {
     interface PosDocument {
         "alt"?: string;
         "onPod-os:init"?: (event: PosDocumentCustomEvent<any>) => void;
+        /**
+          * Indicates that the resource given in `src` property has been loaded.
+         */
+        "onPod-os:resource-loaded"?: (event: PosDocumentCustomEvent<string>) => void;
         "src"?: string;
     }
     interface PosImage {
         "alt"?: string;
         "onPod-os:init"?: (event: PosImageCustomEvent<any>) => void;
+        /**
+          * Indicates that the resource given in `src` property has been loaded.
+         */
+        "onPod-os:resource-loaded"?: (event: PosImageCustomEvent<string>) => void;
         "src"?: string;
     }
     interface PosLabel {
@@ -389,6 +388,10 @@ declare namespace LocalJSX {
     interface PosResource {
         "lazy"?: boolean;
         "onPod-os:init"?: (event: PosResourceCustomEvent<any>) => void;
+        /**
+          * Indicates that the resource given in `uri` property has been loaded.
+         */
+        "onPod-os:resource-loaded"?: (event: PosResourceCustomEvent<any>) => void;
         "uri"?: string;
     }
     interface PosReverseRelations {
@@ -408,8 +411,6 @@ declare namespace LocalJSX {
     }
     interface PosTypeRouter {
         "onPod-os:resource"?: (event: PosTypeRouterCustomEvent<any>) => void;
-    }
-    interface TestComponent {
     }
     interface IntrinsicElements {
         "pos-app": PosApp;
@@ -437,7 +438,6 @@ declare namespace LocalJSX {
         "pos-subjects": PosSubjects;
         "pos-type-badges": PosTypeBadges;
         "pos-type-router": PosTypeRouter;
-        "test-component": TestComponent;
     }
 }
 export { LocalJSX as JSX };
@@ -469,7 +469,6 @@ declare module "@stencil/core" {
             "pos-subjects": LocalJSX.PosSubjects & JSXBase.HTMLAttributes<HTMLPosSubjectsElement>;
             "pos-type-badges": LocalJSX.PosTypeBadges & JSXBase.HTMLAttributes<HTMLPosTypeBadgesElement>;
             "pos-type-router": LocalJSX.PosTypeRouter & JSXBase.HTMLAttributes<HTMLPosTypeRouterElement>;
-            "test-component": LocalJSX.TestComponent & JSXBase.HTMLAttributes<HTMLTestComponentElement>;
         }
     }
 }